requirements.txt
# Simple Spotify API Tool Requirements
# Compatible with Streamlit Cloud

# Core framework
streamlit>=1.28.0

# Data processing
pandas>=1.5.0

# API requests
requests>=2.28.0

# Image handling (for album/artist artwork)
Pillow>=9.0.0

# Excel export functionality
xlsxwriter>=3.0.0
openpyxl>=3.0.0